adding api permission checks

+++ src/modules/fleet/model.js
@@ -1,30 +1,65 @@
 const {executeSQL} = require('../mariadb')
 
-const { getHandle, getOrgTag } = require('../../helpers/db')
+const { getHandle, getOrgTag, getOrgRank } = require('../../helpers/db')
 const { getCitizen } = require ('../citizen/model')
+
+async function canEdit(usr, group) {
+    const user = await getUser(usr)
+    const id = await getID(user.app_metadata.handle)
+    let cmdrs = []
+    if (group.id) {
+        cmdrs = await getCommanders(group.id)
+    } 
+    
+
+    let edit = false
+    
+    switch (group.type) {
+        case 1: // org fleet
+            const rank = await getOrgRank(group.owner, id)
+            if (rank == 5 || cmdrs.includes(user.app_metadata.handle)) {
+                edit = true
+            }
+            break;
+        case 2: // personal fleet
+            if (group.owner === id) {
+                edit = true
+            }
+            break;
+        default:
+            edit = false
+    }
+    return edit
+}
 
 // fleet functions
 
 async function addFleet(usr, data) {
     // TODO: check user CAN add to this owner
-
-    const sql = "INSERT INTO fleet_groups (type, owner, parent, name, purpose) VALUES (?,?,?,?)"
-    const params = [data.type, data.owner, 0, data.name, data.purpose]
-    const res = await executeSQL(sql, params)
-    // TODO: check the sql result
-    return {success: 1, msg: 'Fleet Added!'}
+    if (canEdit(usr, data)) {
+        const sql = "INSERT INTO fleet_groups (type, owner, parent, name, purpose) VALUES (?,?,?,?)"
+        const params = [data.type, data.owner, 0, data.name, data.purpose]
+        const res = await executeSQL(sql, params)
+        // TODO: check the sql result
+        return {success: 1, msg: 'Fleet Added!'}
+    } else {
+        return {success: 0, msg: 'No permissions to add a fleet'}
+    }
 }
 
 async function removeFleet(usr, groupID) {
     // check usr owns org that owns the fleet
+    if (canEdit(usr, await getFleet(groupID))) {
+        // remove all ships in the fleet group
+        await executeSQL('DELETE FROM fleet_ships WHERE parent=?', [groupID])
 
-    // remove all ships in the fleet group
-    await executeSQL('DELETE FROM fleet_ships WHERE parent=?', [groupID])
+        // delete the fleet group
+        await executeSQL('DELETE FROM fleet_groups WHERE id=?', [groupID])
 
-    // delete the fleet group
-    await executeSQL('DELETE FROM fleet_groups WHERE id=?', [groupID])
-
-    return {success: 1, msg: 'Group Removed!'}
+        return {success: 1, msg: 'Group Removed!'}
+    } else {
+        return {success: 0, msg: 'No permission to remove this fleet group'}
+    }
 }
 
 async function getFleet(fleetID) {
@@ -45,10 +80,14 @@ async function getFleet(fleetID) {
 }
 
 async function updateFleet(usr, fleetID, data) {
-    // TODO: Check permission
-    const sql = "UPDATE fleet_groups SET cmdr=? WHERE id=?"
-    await executeSQL(sql, [data.cmdr, fleetID])
-    return {success: 1, msg: 'Commander Updated!'}
+    if (canEdit(usr, await getFleet(fleetID))) {
+        // TODO: Check permission
+        const sql = "UPDATE fleet_groups SET cmdr=? WHERE id=?"
+        await executeSQL(sql, [data.cmdr, fleetID])
+        return {success: 1, msg: 'Commander Updated!'}
+    } else {
+        return {success: 0, msg: 'No permission to update fleet group'}
+    }
 }
 
 async function getGroups(parent) {
@@ -61,9 +100,20 @@ async function getGroups(parent) {
 }
 
 async function addGroup (usr, fleetID, data) {
-    const sql = "INSERT INTO fleet_groups (parent, owner, name, purpose) values (?, ?, ?, ?)"
-    await executeSQL(sql, [fleetID, data.owner, data.name, data.purpose])
-    return {success: 1, msg: 'Group added!'}
+    if (canEdit(usr, await getFleet(fleetID))) {
+        // get parent groups type
+        const rows = await executeSQL("SELECT type FROM fleet_groups WHERE id=?", [fleetID])
+        if (rows.length > 0) {
+            const type = rows[0].type
+            const sql = "INSERT INTO fleet_groups (type, parent, owner, name, purpose) values (?, ?, ?, ?, ?)"
+            await executeSQL(sql, [type, fleetID, data.owner, data.name, data.purpose])
+            return {success: 1, msg: 'Group added!'}
+        } else {
+            return {success: 0, msg: 'Failed adding group. Invalid parent group specified'}
+        }
+    } else {
+        return {success: 0, msg: 'No permissions to add group to this fleet'}
+    }
 }
 
 async function getShips (fleetID) {
@@ -85,16 +135,24 @@ async function getShips (fleetID) {
 }
 
 async function addShip (usr, fleetID, data) {
-    console.log('adding ship', data)
-    const sql = "INSERT INTO fleet_ships (fleet, parent, ship) values (?, ?, ?)"
-    await executeSQL(sql, [fleetID, data.group, data.ship])
-    return {success: 1, msg: 'Ship added!'}
+    if (canEdit(usr, await getFleet(data.group))) {
+        console.log('adding ship', data)
+        const sql = "INSERT INTO fleet_ships (fleet, parent, ship) values (?, ?, ?)"
+        await executeSQL(sql, [fleetID, data.group, data.ship])
+        return {success: 1, msg: 'Ship added!'}
+    } else {
+        return {success: 0, msg: 'No permissions to add a ship to this fleet group'}
+    }
 }
 
 async function removeShip (usr, fleetID, shipID) {
-    const sql = "DELETE FROM fleet_ships WHERE fleet=? AND ship=?"
-    await executeSQL(sql, [fleetID, shipID])
-    return {success: 1, msg: 'Ship removed!'}
+    if (canEdit(usr, await getFleet(fleetID))) {
+        const sql = "DELETE FROM fleet_ships WHERE fleet=? AND ship=?"
+        await executeSQL(sql, [fleetID, shipID])
+        return {success: 1, msg: 'Ship removed!'}
+    } else {
+        return {success: 0, msg: 'No permission to remove ship from this fleet group'}
+    }
 }
 
 // crew functions
@@ -106,15 +164,23 @@ async function getCrew(fleetID, shipID) {
 }
 
 async function addCrew(usr, fleetID, shipID, data) {
-    // add a crewmen to the specified fleet ship
-    await executeSQL('INSERT INTO fleet_personnel (fleet, ship, citizen, role) VALUES (?, ?, ?, ?)', [fleetID, shipID, data.handle, data.role])
-    return {success: 1, msg: 'Successfully added crewmember!'}
+    if (canEdit(usr, await getFleet(fleetID))) {
+        // add a crewmen to the specified fleet ship
+        await executeSQL('INSERT INTO fleet_personnel (fleet, ship, citizen, role) VALUES (?, ?, ?, ?)', [fleetID, shipID, data.handle, data.role])
+        return {success: 1, msg: 'Successfully added crewmember!'}
+    } else {
+        return {success: 0, msg: 'No permission to edit the crew of this fleet ship'}
+    }
 }
 
 async function removeCrew(usr, fleetID, shipID, crewID) {
-    // remove the specified crewmember
-    await executeSQL('DELETE FROM fleet_personnel WHERE fleet=? AND ship=? AND citizen=?', [fleetID, shipID, crewID])
-    return {success: 1, msg: 'Successfully removed crewmember!'}
+    if (canEdit(usr, await getFleet(fleetID))) {
+        // remove the specified crewmember
+        await executeSQL('DELETE FROM fleet_personnel WHERE fleet=? AND ship=? AND citizen=?', [fleetID, shipID, crewID])
+        return {success: 1, msg: 'Successfully removed crewmember!'}
+    } else {
+        return {success: 0, msg: 'No permission to edit the crew of this fleet ship'}
+    }
 }
 
 async function getCommanders(fleetID) {
@@ -129,7 +195,7 @@ async function getCommanders(fleetID) {
         if (data.parent !== 0) { // fleet root group
             commanders = commanders.concat(await getCommanders(data.parent))
         }
-        
+
         return commanders
     } else {
         return []
